delete w comments

+++ src/app/api/posts/edit/[slug]/route.js
@@ -23,6 +23,7 @@ export const GET = async (req, { params }) => {
 
 // UPDATE SINGLE POST By ID
 export const POST = async (req, { params }) => {
+  if (req.method !== 'POST') return
   console.log('method', req.method)
   const id = params.slug
   console.log(id)
@@ -55,11 +56,37 @@ export const POST = async (req, { params }) => {
 
 // DELETE SINGLE POST By ID
 export const DELETE = async (req, { params }) => {
+  if (req.method !== 'DELETE') return
+
   console.log('method', req.method)
   const id = params.slug
 
   try {
-    const post = await prisma.post.delete({
+    const post = await prisma.post.findUnique({
+      where: {
+        id: id,
+      },
+    })
+
+    if (post) {
+      const postSlugDB = post.slug
+      const comments = await prisma.comment.findMany({
+        where: {
+          postSlug: postSlugDB,
+        },
+      })
+      if (comments) {
+        console.log(comments)
+        await prisma.comment.deleteMany({
+          where: {
+            postSlug: postSlugDB,
+          },
+        })
+      } else {
+        console.log('no comments')
+      }
+    }
+    await prisma.post.delete({
       where: {
         id: id,
       },
